requirements.txt
# Python 3.10.11
emotiefflib==1.1.1
faster-whisper==1.2.0
mediapipe
numpy>=1.24.0,<2.0.0
openai[realtime]==2.6.0
PyAudio==0.2.14
python-dotenv==1.1.1
SpeechRecognition==3.14.3
timm==1.0.20
